+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Added
+
+- Added support for the [PostgreSQL connector](https://trino.io/docs/current/connector/postgresql.html) using the new generic database connection mechanism.
+  Previously, users had to use the `generic` connector ([#883]).
+
 ### Changed
 
 - BREAKING: `configOverrides` now only accepts the supported config file names
@@ -26,6 +31,7 @@ All notable changes to this project will be documented in this file.
 [#869]: https://github.com/stackabletech/trino-operator/pull/869
 [#876]: https://github.com/stackabletech/trino-operator/pull/876
 [#878]: https://github.com/stackabletech/trino-operator/pull/878
+[#883]: https://github.com/stackabletech/trino-operator/pull/883
 
 ## [26.3.0] - 2026-03-16